requirements-deploy.txt
# Core web stack
Django>=5.0,<5.1
djangorestframework>=3.14,<3.16
django-cors-headers>=4.3,<4.5
django-environ>=0.11,<0.12
gunicorn>=21.2,<21.3

# Data and storage
psycopg2-binary>=2.9,<2.10
redis>=5.0,<6.0
supabase>=2.4,<3.0

# Runtime services
python-dotenv>=1.0.0,<1.1.0
boto3>=1.34,<1.36
requests>=2.31,<2.33
whitenoise[brotli]>=6.6.0

# Image and AI runtime
Pillow>=10.2.0
PyMuPDF>=1.27,<1.28
opencv-python-headless>=4.8,<4.14
numpy>=1.26,<2.0
fastapi>=0.111,<0.116
uvicorn>=0.30,<0.31
python-multipart>=0.0.9,<0.1
SQLAlchemy>=2.0,<2.1
openai>=2.26,<3.0
langchain-core>=1.2,<1.3
langchain-openai>=1.1,<1.2
langchain-chroma==0.2.3
chromadb>=0.5,<0.7

# Security and API docs
python-jose[cryptography]>=3.3,<3.4
passlib[bcrypt]>=1.7,<1.8
drf-spectacular>=0.27,<0.29

# Background jobs and feed refresh
apscheduler==3.10.4
beautifulsoup4>=4.12,<5.0
